update mcp pipeline
